Auth ve Json Operations iyilşetirildi. Auth için void metodlar User dönecek şekilde güncellendi. Json da ise read file ın null dönmesi ihtimali kontrol altına alındı.

+++ Readed Book Lister/App Logics/JsonOperations.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using Readed_Book_Lister.Constants;
 using Readed_Book_Lister.Entities;
 using System;
 using System.Collections.Generic;
@@ -11,7 +12,7 @@ namespace Readed_Book_Lister.Methods.Helpers
 {
     public class JsonOperations
     {
-        // Deletelerde ilgili managerlerin getall ları null ise de kullanılabilir böyle kalsa da olur. kontrol ok.
+        // Checked.
 
         const string userFileName = "users.json";
         const string userBooksFileName = "userbooks.json";
@@ -35,24 +36,28 @@ namespace Readed_Book_Lister.Methods.Helpers
         {
             var jsonUserList = File.ReadAllText(userFileName);
             var userList = JsonConvert.DeserializeObject<List<User>>(jsonUserList);
-            if (userList.Count == 0)
+            if (userList != null)
             {
-
-                File.Delete(userFileName);
+                if (userList.Count == 0)
+                {
+                    File.Delete(userFileName);
+                }
             }
-
+            System.Windows.Forms.MessageBox.Show(Messages.UserFileNotExist);
         }
 
         public static void DeleteUserBooksFile()
         {
             var jsonUserBookList = File.ReadAllText(userBooksFileName);
             var userBooksList = JsonConvert.DeserializeObject<List<User>>(jsonUserBookList);
-            if (userBooksList.Count == 0)
+            if (userBooksList != null)
             {
-                File.Delete(userBooksFileName);
+                if (userBooksList.Count == 0)
+                {
+                    File.Delete(userBooksFileName);
+                }
             }
-
+            System.Windows.Forms.MessageBox.Show(Messages.UserBooksFileNotExist);
         }
-
     }
 }
